Add files via upload
This project demonstrates image classification using the YOLO11 model on a custom dataset of flowers—specifically Daisy and Dandelion images. Implemented in Google Colab, this project leverages Ultralytics YOLO11 for training and evaluation, showcasing how YOLO can be used not only for object detection but also for image classification tasks.